dried up the code

+++ server.py
@@ -13,14 +13,13 @@ logger = logging.getLogger(__name__)
 async def create_zip_archive(directory, delay, photo_path):
     zip_bytes = b''
 
+    process = await asyncio.create_subprocess_exec(
+        'zip', '-r', '-', f'{photo_path}/{directory}',
+        stdout=asyncio.subprocess.PIPE,
+        stderr=asyncio.subprocess.PIPE,
+        cwd='.'
+    )
     try:
-        process = await asyncio.create_subprocess_exec(
-            'zip', '-r', '-', f'{photo_path}/{directory}',
-            stdout=asyncio.subprocess.PIPE,
-            stderr=asyncio.subprocess.PIPE,
-            cwd='.'
-        )
-    
         while True:
             chunk = await process.stdout.read(500 * 1024)
             if not chunk:
@@ -33,10 +32,10 @@ async def create_zip_archive(directory, delay, photo_path):
     except asyncio.CancelledError:
         logger.debug(f"Download was interrupted")
         raise
-    except (KeyboardInterrupt, IndexError, BaseException):
-        process.kill()
     finally:
-        process.kill()
+        if process.returncode != 0:
+            process.kill()
+            await process.communicate()
 
     return zip_bytes
